% Generated by roxygen2 (4.0.2): do not edit by hand
\name{df_to_string}
\alias{df_to_string}
\title{Format a df of addresses for upload}
\usage{
df_to_string(addresses_df)
}
\arguments{
\item{addresses}{A df with two columns, an id and address character strings to be geocoded}
}
\description{
Format a df of addresses for upload
}
\examples{
\dontrun{
data(chicago_landmarks)
addresses <- chicago_landmarks[,"Address"]
addresses <- paste(addresses, "chicago IL")
address_str <- vec_to_string(addresses)
}
df_to_string
}
\keyword{batch}
\keyword{geocode}

