% Generated by roxygen2 (4.0.2): do not edit by hand
\name{format_vec_for_upload}
\alias{format_vec_for_upload}
\title{Format a vector of addresses for upload}
\usage{
format_vec_for_upload(addresses)
}
\arguments{
\item{addresses}{A vector of character strings to be geocoded}
}
\description{
Format a vector of addresses for upload
}
\examples{
\dontrun{
load("chicago_landmarks.Rdata")
addresses <- chicago_landmarks[,"Address"]
addresses <- paste(addresses, "chicago IL")
big_string <- format_vec_for_upload(addresses)
}
format_vec_for_upload
}
\keyword{batch}
\keyword{geocode}

