% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geocodeHERE_batch_status}
\alias{geocodeHERE_batch_status}
\title{Check the status on a batch geocoding job}
\usage{
geocodeHERE_batch_status(request_id = "", full_list = FALSE, App_id = "",
  App_code = "")
}
\arguments{
\item{request_id}{Character string containing a request_id. This is returned from geocodeHERE_batch_upload(...)}

\item{full_list}{TRUE / FALSE indicating whether to return the full response from Nokia HERE or just the "Status" portion of the response}

\item{App_id}{App_id to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}

\item{App_code}{App_code to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}
}
\description{
Check the status on a batch geocoding job
}
\examples{
\dontrun{
request_id <- geocodeHERE_batch_upload(file_name = "example.txt", email_address = "youremail<at>domain.com")
geocodeHERE_batch_status(request_id)
}
geocodeHERE_batch_status
}
\keyword{batch}
\keyword{geocode}

