% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_IDs}
\alias{get_IDs}
\title{Read in App_id and App_code}
\usage{
get_IDs(file_name = "R/nokia_HERE.txt")
}
\arguments{
\item{file_name}{The location of the file}
}
\description{
Loads a file that contains the App_id and App_code
}
\examples{
\dontrun{
ids = get_IDs(file_name="~/nokia_HERE.txt")
geocodeHERE_simple("chicago")
geocodeHERE_simple("wrigley field chicago IL")
geocodeHERE_simple("233 S Wacker Dr, Chicago, IL 60606")
}
get_IDs
}
\keyword{App_code}
\keyword{App_id}

