% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{chicago_landmarks}
\alias{chicago_landmarks}
\title{Chicago Landmarks.}
\format{Data frame with columns}
\source{
http://en.wikipedia.org/wiki/List_of_Chicago_Landmarks
}
\usage{
chicago_landmarks
}
\description{
Addresses and other data from Chicago Landmarks
}
\examples{
chicago_landmarks
}
\keyword{datasets}

