% Generated by roxygen2 (4.0.2): do not edit by hand
\name{df_to_string}
\alias{df_to_string}
\title{Format a df of addresses for upload}
\usage{
df_to_string(addresses_df)
}
\arguments{
\item{addresses}{A df with two columns, a unique id and address character strings to be geocoded. Only results with valid lat, lng are returned, so the unique id is used to match back to the data later on.}
}
\value{
a long string consisting of all of the addresses to be geocoded formatted for the POST request in geocodeHERE_batch_upload()
}
\description{
Format a df of addresses for upload
}
\examples{
\dontrun{
data(chicago_landmarks)
addresses <- chicago_landmarks[,"Address"]
addresses <- paste(addresses, "chicago IL")
address_str <- vec_to_string(addresses)
}
}
\keyword{batch}
\keyword{geocode}

