% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geocodeHERE_batch_status}
\alias{geocodeHERE_batch_status}
\title{Check the status on a batch geocoding job}
\usage{
geocodeHERE_batch_status(request_id = "", full_list = FALSE, App_id = "",
  App_code = "")
}
\arguments{
\item{request_id}{Character string containing a request_id. This is returned from geocodeHERE_batch_upload(...)}

\item{full_list}{TRUE / FALSE indicating whether to return the full response from Nokia HERE or just the "Status" portion of the response}

\item{App_id}{App_id to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}

\item{App_code}{App_code to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}
}
\description{
Check the status on a batch geocoding job
}
\examples{
\dontrun{
addresses <- chicago_landmarks[,"Address"]
addresses <- paste(addresses, "chicago IL")
addresses_df <- data.frame(id=1:length(addresses), addresses=addresses)
address_str <- df_to_string(addresses_df)
request_id <- geocodeHERE_batch_upload(address_string = address_str, email_address = "youremail<at>domain.com")
geocodeHERE_batch_status(request_id)
geocode_data <- geocodeHERE_batch_get_data(request_id)
addresses_df <- merge(addresses_df, geocode_data, by.x="id", by.y="recId", all.x=T)
}
geocodeHERE_batch_status
}
\keyword{batch}
\keyword{geocode}

