% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geocodeHERE_batch_upload}
\alias{geocodeHERE_batch_upload}
\title{Upload a file to Nokia HERE batch geocoding API}
\usage{
geocodeHERE_batch_upload(address_string, email_address, App_id = "",
  App_code = "", quiet = TRUE)
}
\arguments{
\item{address_string}{Character string containing the addresses to be geocoded. Output from 'format_vec_for_upload(...)}

\item{email_address}{Character string containing an email address. Nokia will email you here when the job is done.}

\item{App_id}{App_id to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}

\item{App_code}{App_code to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}

\item{quiet}{TRUE / FALSE indicating whether to write the POST information to the console}
}
\description{
Upload a file to Nokia HERE batch geocoding API
}
\examples{
\dontrun{
#' data(chicago_landmarks)
addresses <- chicago_landmarks[,"Address"]
addresses <- paste(addresses, "chicago IL")
address_str <- format_vec_for_upload(addresses)
geocodeHERE_batch_upload(address_string = address_str, email_address = "youremail<at>domain.com")
}
geocodeHERE_batch_upload
}
\keyword{batch}
\keyword{geocode}

