% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geocodeHERE_batch_upload}
\alias{geocodeHERE_batch_upload}
\title{Upload a file to Nokia HERE batch geocoding API}
\usage{
geocodeHERE_batch_upload(file_name, email_address, App_id = "",
  App_code = "", quiet = TRUE)
}
\arguments{
\item{file_name}{Character string containing the path to the file to be uploaded}

\item{email_address}{Character string containing an email address. Nokia will email you here when the job is done.}

\item{App_id}{App_id to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}

\item{App_code}{App_code to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}

\item{quiet}{TRUE / FALSE indicating whether to write the POST information to the console}
}
\description{
Upload a file to Nokia HERE batch geocoding API
}
\examples{
\dontrun{
geocodeHERE_batch_upload(file_name = "example.txt", email_address = "youremail<at>domain.com")
}
geocodeHERE_batch_upload
}
\keyword{batch}
\keyword{geocode}

