% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geocodeHERE_simple}
\alias{geocodeHERE_simple}
\title{Attempt to geocode a string}
\usage{
geocodeHERE_simple(search, App_id = "", App_code = "")
}
\arguments{
\item{search}{A string to search}

\item{App_id}{App_id to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}

\item{App_code}{App_code to use the production HERE API. Get one here... http://developer.here.com/get-started. If left blank, will default to demo key with an unknown usage limit.}
}
\value{
A list containing Latitude and Longitude if found, NA otherwise
}
\description{
Enter a string and if found, the latitude and longitude is returned using the HERE API
}
\examples{
\dontrun{
geocodeHERE_simple("chicago")
geocodeHERE_simple("wrigley field chicago IL")
geocodeHERE_simple("233 S Wacker Dr, Chicago, IL 60606")
}
}
\keyword{geocode}

