% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geocodeHERE_simple}
\alias{geocodeHERE_simple}
\title{Attempt to geocode a string}
\usage{
geocodeHERE_simple(search, format = "json", ids = ids)
}
\arguments{
\item{search}{A string to search}

\item{format}{json or xml. Defaults to json.}

\item{ids}{The output from the get_IDs function. This contains the App_id and App_code required for authentication by the HERE API}
}
\description{
Enter a string and have latitude and longitude returned using the HERE API
}
\examples{
\dontrun{
ids = get_IDs(file_name="~/nokia_HERE.txt")
geocodeHERE_simple("chicago")
geocodeHERE_simple("wrigley field chicago IL")
geocodeHERE_simple("233 S Wacker Dr, Chicago, IL 60606")
}
geocodeHERE_simple
}
\keyword{geocode}

